done favourites

+++ src/Controller/FavouritesController.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace Controller;
 
 use Model\Product;
@@ -21,34 +22,34 @@ class FavouritesController {
         }
 
         $userId = $_SESSION['userId'];
-        $productId = $_POST['productId'];
+        $productId = $_POST['productId'] ?? null;
         $amount = $_POST['amount'] ?? 1;
 
         if ($productId === null) {
-            echo "Product ID is missing!";
-            exit();
-        }
-        if ($amount === null) {
-            echo "Amount is missing!";
+            echo "Product id is missing!";
             exit();
         }
 
         $price = $this->product->getByProductId((int)$productId);
-        $isProductInFavourites = $this->favourites->getByUserIdAndProductId($userId, (int)$productId);
-
-        if ($isProductInFavourites === false) {
-            $this->favourites->addProductToFavourites($userId, (int)$productId, (int)$amount, $price['price']);
-            header('Location: /favourites');
-            exit();
-        } else {
-            $newAmount = $amount + $isProductInFavourites['amount'];
-            $this->favourites->plusProductAmountInCart($userId, (int)$productId, $newAmount);
-            header('Location: /favourites');
+        if (!$price) {
+            echo "Invalid product ID!";
             exit();
         }
+
+        $isProductInFavourites = $this->favourites->getByUserIdAndProductId($userId, (int)$productId);
+
+        if (!$isProductInFavourites) {
+            $this->favourites->addProductToFavourites($userId, (int)$productId, (int)$amount, $price['price']);
+        } else {
+            $newAmount = $amount + $isProductInFavourites['amount'];
+            $this->favourites->updateProductAmountInCart($userId, (int)$productId, $newAmount);
+        }
+
+        header('Location: /favourites');
+        exit();
     }
 
-    public function lookFavourites() {
+    public function getFavourites() {
         session_start();
         if (!isset($_SESSION['userId'])) {
             header('Location: /login');
@@ -58,6 +59,7 @@ class FavouritesController {
         $userId = $_SESSION['userId'];
         $favouritesProducts = $this->favourites->getFavouritesByUserId($userId);
         $productsInFavourites = [];
+
         foreach ($favouritesProducts as $elem) {
             $product = $this->product->getByProductId((int)$elem['product_id']);
             if ($product) {
@@ -77,13 +79,20 @@ class FavouritesController {
         }
 
         $userId = $_SESSION['userId'];
-        $productId = $_POST['product-id'];
+        $productId = $_POST['product-id'] ?? null;
+
+        if ($productId === null) {
+            echo "Product id is missing!";
+            exit();
+        }
+
         $isProductInFavourites = $this->favourites->getByUserIdAndProductId($userId, (int)$productId);
 
         if ($isProductInFavourites) {
             $this->favourites->deleteProduct($userId, (int)$productId);
-            header('Location: /favourites');
-            exit();
         }
+
+        header('Location: /favourites');
+        exit();
     }
 }
